requirements.txt
pydantic==1.10.7
pyyaml==6.0
colorama==0.4.6
urllib3==1.26.15
httplib2==0.22.0
pandas==1.5.3
numpy==1.24.2
requests==2.28.2
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
faker==18.3.1
